use double instead of float as main numerical type in FXGLMath

+++ fxgl-samples/src/main/java/s09advanced/NoiseSample.java
@@ -68,7 +68,7 @@ public class NoiseSample extends GameApplication {
 
     @Override
     protected void onUpdate(double tpf) {
-        float n = FXGLMath.noise1D(t);
+        double n = FXGLMath.noise1D(t);
 
         // flickering rect
         rect.setFill(Color.color(1.0, n, n));
